Remove redundant code from MapViewController
